mcp_query_handler/src/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy[asyncio]
psycopg2-binary
neo4j
openai
